Speech-Modeling-master/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
pandas
librosa
pyyaml
webrtcvad
scipy
numpy
colorama
tqdm
h5py
matplotlib
soundfile
tensorboard